requirements.txt
streamlit
pandas
matplotlib
seaborn
openpyxl
Pillow